+++ frontend/src/features/material-details/components/DeliveryInformationModal.tsx
@@ -197,7 +197,7 @@ const createDeliveryColumns = (handleDelete: (row: Delivery) => void) => {
                 return (
                     <Box display="flex" textAlign="center" alignItems="center" justifyContent="center" width="100%" height="100%">
                         {!data.row.reported && (
-                            <Button variant="text" color="error" onClick={() => handleDelete(data.row)}>
+                            <Button variant="text" color="error" onClick={() => handleDelete(data.row)} data-testid="delete-delivery">
                                 <Delete></Delete>
                             </Button>
                         )}
@@ -313,7 +313,7 @@ export const DeliveryInformationModal = ({
     }, [delivery]);
     return (
         <>
-            <Dialog open={open && delivery !== null} onClose={handleClose}>
+            <Dialog open={open && delivery !== null} onClose={handleClose} data-testid="delivery-modal">
                 <DialogTitle variant="h3" textAlign="center">
                     {capitalize(mode)} Delivery Information
                 </DialogTitle>
@@ -324,7 +324,7 @@ export const DeliveryInformationModal = ({
                                 <GridItem label="Material Number" value={temporaryDelivery.ownMaterialNumber ?? ''} />
                                 <Grid item xs={6}>
                                     <LabelledAutoComplete
-                                        id="partnerBpns"
+                                        id="ownBpns"
                                         options={sites ?? []}
                                         getOptionLabel={(option) => option.name ?? ''}
                                         isOptionEqualToValue={(option, value) => option?.bpns === value.bpns}
@@ -350,6 +350,7 @@ export const DeliveryInformationModal = ({
                                             formError &&
                                             (direction === DirectionType.Outbound ? !temporaryDelivery.originBpns : !temporaryDelivery.destinationBpns)
                                         }
+                                        data-testid="delivery-own-bpns-field"
                                     />
                                 </Grid>
                                 <Grid item xs={6}>
@@ -365,6 +366,7 @@ export const DeliveryInformationModal = ({
                                         label="Departure Type*"
                                         placeholder="Select the type of departure"
                                         error={formError && !temporaryDelivery?.departureType}
+                                        data-testid="delivery-departure-type-field"
                                     />
                                 </Grid>
                                 <Grid item xs={6}>
@@ -385,9 +387,10 @@ export const DeliveryInformationModal = ({
                                                 (temporaryDelivery?.arrivalType === 'actual-arrival' &&
                                                     temporaryDelivery?.departureType !== 'actual-departure'))
                                         }
+                                        data-testid="delivery-arrival-type-field"
                                     ></LabelledAutoComplete>
                                 </Grid>
-                                <Grid item xs={6} display="flex" alignItems="end">
+                                <Grid item xs={6} display="flex" alignItems="end" data-testid="delivery-departure-time-field">
                                     <DateTime
                                         label="Departure Time*"
                                         placeholder="Pick Departure Date"
@@ -406,7 +409,7 @@ export const DeliveryInformationModal = ({
                                         }
                                     />
                                 </Grid>
-                                <Grid item xs={6} display="flex" alignItems="end">
+                                <Grid item xs={6} display="flex" alignItems="end" data-testid="delivery-arrival-time-field">
                                     <DateTime
                                         label="Arrival Time*"
                                         placeholder="Pick Arrival Date"
@@ -437,6 +440,7 @@ export const DeliveryInformationModal = ({
                                         }
                                         value={partners?.find((p) => p.bpnl === temporaryDelivery.partnerBpnl) ?? null}
                                         isOptionEqualToValue={(option, value) => option?.bpnl === value?.bpnl}
+                                        data-testid="delivery-partner-field"
                                     />
                                 </Grid>
                                 <Grid item xs={6}>
@@ -470,6 +474,7 @@ export const DeliveryInformationModal = ({
                                             formError &&
                                             (direction === DirectionType.Inbound ? !temporaryDelivery.originBpns : !temporaryDelivery.destinationBpns)
                                         }
+                                        data-testid="delivery-partner-bpns-field"
                                     />
                                 </Grid>
                                 <Grid item xs={6}>
@@ -486,6 +491,7 @@ export const DeliveryInformationModal = ({
                                             }))
                                         }
                                         sx={{ marginTop: '.5rem' }}
+                                        data-testid="delivery-quantity-field"
                                     />
                                 </Grid>
                                 <Grid item xs={6}>
@@ -506,6 +512,7 @@ export const DeliveryInformationModal = ({
                                         error={formError && !temporaryDelivery?.measurementUnit}
                                         onChange={(_, value) => setTemporaryDelivery((curr) => ({ ...curr, measurementUnit: value?.key }))}
                                         isOptionEqualToValue={(option, value) => option?.key === value?.key}
+                                        data-testid="delivery-uom-field"
                                     />
                                 </Grid>
                                 <Grid item xs={6}>
@@ -519,6 +526,7 @@ export const DeliveryInformationModal = ({
                                             setTemporaryDelivery({ ...temporaryDelivery, trackingNumber: event.target.value })
                                         }
                                         sx={{ marginTop: '.5rem' }}
+                                        data-testid="delivery-tracking-number-field"
                                     />
                                 </Grid>
                                 <Grid item xs={6}>
@@ -538,6 +546,7 @@ export const DeliveryInformationModal = ({
                                         error={formError && !temporaryDelivery?.incoterm}
                                         onChange={(_, value) => setTemporaryDelivery((curr) => ({ ...curr, incoterm: value?.key }))}
                                         isOptionEqualToValue={(option, value) => option?.key === value?.key}
+                                        data-testid="delivery-incoterm-field"
                                     />
                                 </Grid>
                                 <Grid item xs={6}>
@@ -550,6 +559,7 @@ export const DeliveryInformationModal = ({
                                         onChange={(event) =>
                                             setTemporaryDelivery({ ...temporaryDelivery, customerOrderNumber: event.target.value })
                                         }
+                                        data-testid="delivery-customer-order-number-field"
                                     />
                                 </Grid>
                                 <Grid item xs={6}>
@@ -565,6 +575,7 @@ export const DeliveryInformationModal = ({
                                                 customerOrderPositionNumber: event.target.value,
                                             })
                                         }
+                                        data-testid="delivery-customer-order-position-field"
                                     />
                                 </Grid>
                                 <Grid item xs={6}>
@@ -579,6 +590,7 @@ export const DeliveryInformationModal = ({
                                         onChange={(event) =>
                                             setTemporaryDelivery({ ...temporaryDelivery, supplierOrderNumber: event.target.value })
                                         }
+                                        data-testid="delivery-supplier-order-number-field"
                                     />
                                 </Grid>
                             </>
@@ -621,6 +633,7 @@ export const DeliveryInformationModal = ({
                                 color="primary"
                                 sx={{ display: 'flex', gap: '.25rem' }}
                                 onClick={() => handleSaveClick()}
+                                data-testid="save-delivery-button"
                             >
                                 <Save></Save> Save
                             </Button>