+++ frontend/src/features/material-details/components/SummaryPanel.tsx
@@ -29,6 +29,8 @@ import { DirectionType } from '@models/types/erp/directionType';
 import { Supply } from '@models/types/data/supply';
 import { useReportedDaysOfSupply } from '../hooks/useReportedDaysOfSupply';
 
+type DataAttributes = Record<`data-${string}`, string>;
+
 type SummaryPanelProps<TType extends SummaryType> = {
     sx?: SxProps<Theme>
     title?: string;
@@ -42,12 +44,12 @@ type SummaryPanelProps<TType extends SummaryType> = {
 
 export function OwnSummaryPanel<TType extends SummaryType>({ summary, materialNumber, site, partnerBpnl, ...props }: SummaryPanelProps<TType>) {
     const { supplies } = useDaysOfSupply(materialNumber, summary.type === 'demand' ? DirectionType.Inbound : DirectionType.Outbound, site, partnerBpnl);
-    return <SummaryPanelContent summary={summary} {...props} supplies={supplies ?? []} />
+    return <SummaryPanelContent summary={summary} {...props} supplies={supplies ?? []} data-testid="own-summary-panel" />
 }
 
 export function ReportedSummaryPanel<TType extends SummaryType>({ summary, materialNumber, site, partnerBpnl, ...props }: SummaryPanelProps<TType>) {
     const { supplies } = useReportedDaysOfSupply(materialNumber, summary.type === 'demand' ? DirectionType.Inbound : DirectionType.Outbound, site, partnerBpnl);
-    return <SummaryPanelContent summary={summary} {...props} supplies={supplies ?? []} />
+    return <SummaryPanelContent summary={summary} {...props} supplies={supplies ?? []} data-testid={`reported-summary-panel-${site ? site : partnerBpnl}`}/>
 }
 
 type SummaryPanelContentProps<TType extends SummaryType> = {
@@ -56,13 +58,13 @@ type SummaryPanelContentProps<TType extends SummaryType> = {
     summary: Summary<TType>;
     showHeader?: boolean;
     includeDaysOfSupply?: boolean;
-};
+} & DataAttributes;
 
-function SummaryPanelContent<TType extends SummaryType>({ sx = {}, title, summary, supplies, showHeader = false, includeDaysOfSupply = false }: SummaryPanelContentProps<TType> & { supplies: Supply[]}) {
+function SummaryPanelContent<TType extends SummaryType>({ sx = {}, title, summary, supplies, showHeader = false, includeDaysOfSupply = false, ...props}: SummaryPanelContentProps<TType> & { supplies: Supply[]}) {
     const theme = useTheme();
     const { calendarWeeks, expandWeek } = useCalendarWeeks();
     return (
-        <Stack direction="row" sx={sx}>
+        <Stack direction="row" sx={sx} {...props}>
             <Stack
                 flex={1}
                 minWidth="12rem"